requirements.txt
azure-identity==1.19.0
azure-core==1.32.0
